Cleanup logging, fix a few minor problems, swap order of metadata and data uploads so they fail correctly.

+++ src/pipelines/canu/Grid_Cloud.pm
@@ -1,4 +1,4 @@
-
+     
 ###############################################################################
  #
  #  This file is part of canu, a software program that assembles whole-genome
@@ -66,7 +66,10 @@ use canu::Execution;
 
 #use canu::Grid "formatAllowedResources";
 
-my $showWork = 1;
+my $showWork   = 1;
+
+my $retryCount = 5;    #  Try a few times to upload the file.
+my $retryDelay = 10;   #  Wait a little bit before retrying.
 
 
 #  Convert a/path/to/file to ../../../..
@@ -203,7 +206,7 @@ sub renameStashedFile ($$) {
     my $pr      = getGlobal("objectStoreProject");
 
     if (isOS() eq "DNANEXUS") {
-        print STDERR "renameStashedFile()-- \"$ns/$oldname\" -> \"$ns/$newname\"\n"   if ($showWork);
+        print STDERR "-- Rename storage object '$pr:$ns/$oldname' -> '$pr:$ns/$newname'\n"   if ($showWork);
 
         if (runCommandSilently(".", "$dx mv \"$pr:$ns/$oldname\" \"$pr:$ns/$newname\"", 1)) {
             caExit("failed to rename object store file", undef);
@@ -222,7 +225,7 @@ sub removeStashedFile ($) {
     return   if (! fileExists("$name", 1));
 
     if (isOS() eq "DNANEXUS") {
-        print STDERR "removeStashedFile()-- $ns/$name'\n"   if ($showWork);
+        print STDERR "-- Remove storage object '$pr:$ns/$name'\n"   if ($showWork);
 
         if (runCommandSilently(".", "$dx rm --recursive \"$pr:$ns/$name\"", 1)) {
             caExit("failed to remove object store file", undef);
@@ -242,23 +245,28 @@ sub stashFile ($) {
     my $ns       = getGlobal("objectStoreNameSpace");
     my $pr       = getGlobal("objectStoreProject");
 
-    my $retries  = 5;    #  Try a few times to upload the file.   (also set in stashFileShellCode())
-    my $delay    = 10;   #  Wait a little bit before retrying.
-
     return   if (! -e $pathname);
 
-    if (isOS() eq "DNANEXUS") {
-        print STDERR "stashFile()-- '$pathname' to project '$pr' namespace '$ns' path '$path' name '$name'.\n"   if ($showWork);
+    my $retries  = $retryCount;    #  Try a few times to upload the file.
+    my $delay    = $retryDelay;    #  Wait a little bit before retrying.
 
-        if (runCommandSilently(".", "$dx rm --recursive \"$pr:$ns/$path/$name\"", 1)) {
-            caExit("failed to remove object store file", undef);
+    if (isOS() eq "DNANEXUS") {
+        print STDERR "-- Store '$pathname' to storage object '$pr:$ns/$path/$name'.\n"   if ($showWork);
+
+        #  If the filename exists, remove it before uploading.
+        if (fileExists($pathname, 1)) {
+            if (runCommandSilently(".", "$dx rm --recursive \"$pr:$ns/$path/$name\"", 1)) {
+                caExit("failed to remove object store file", undef);
+            }
         }
 
         #  Try a couple of times to upload the file.  If the UA fails, delay a bit and retry.
         while (($retries > 0) &&
                (runCommandSilently(".", "$ua --do-not-compress --wait-on-close --project \"$pr\" --folder \"$ns/$path/\" --name \"$name\" \"$pathname\"", 0))) { 
             $retries--;
-            print STDERR "stashFile()-- Failed to stash file '$pathname', wait $delay seconds and try again ($retries times left).\n";
+            print STDERR "-- WARNING:\n";
+            print STDERR "-- WARNING:  Failed to store '$pathname'.  Wait $delay seconds and try again ($retries times left).\n";
+            print STDERR "-- WARNING:\n";
             sleep($delay);
         }
 
@@ -283,7 +291,7 @@ sub fetchFile ($) {
     return   if (! fileExists($pathname, 1));   #  If it doesn't exist in the store, we don't fetch it either.  Because it doesn't exist.
 
     if (isOS() eq "DNANEXUS") {
-        print STDERR "fetchFile()-- from project '$pr' path '$path' name '$name' -> '$pathname'\n"   if ($showWork);
+        print STDERR "-- Fetch '$pathname' from storage object '$pr:$path/$name'.\n"   if ($showWork);
         make_path($path);
 
         if (runCommandSilently(".", "$da download --output \"$pathname\" \"$pr:$ns/$path/$name\"", 1)) {
@@ -308,26 +316,25 @@ sub stashFileShellCode ($$$) {
     my $pr     = getGlobal("objectStoreProject");
     my $code   = "";
 
-    my $retries  = 5;    #  Try a few times to upload the file.   (also set in stashFile())
-    my $delay    = 10;   #  Wait a little bit before retrying.
-
     if (isOS() eq "DNANEXUS") {
         $code .= "\n";
         $code .= "${indent}if [ -e \"./$file\" ] ; then\n";
         $code .= "${indent}  fold=`dirname \"$path/$file\"`\n";
         $code .= "${indent}  name=`basename \"$file\"`\n";
         $code .= "${indent}\n";
-        $code .= "${indent}  $dx rm --recursive \"$pr:$ns/$path/$file\"\n";
+        $code .= "${indent}  if $dx describe --name \"$pr:$ns/$path/$file\" > /dev/null 2>&1 ; then\n";
+        $code .= "${indent}    $dx rm --recursive \"$pr:$ns/$path/$file\"\n";
+        $code .= "${indent}  fi\n";
         $code .= "${indent}\n";
-        $code .= "${indent}  retries=$retries\n";
+        $code .= "${indent}  retries=$retryCount\n";
         $code .= "${indent}  while [ \$retries -gt 0 ] && \\\n";
         $code .= "${indent}        ! $ua --do-not-compress --wait-on-close --project \"$pr\" --folder \"$ns/\$fold/\" --name \"\$name\" \"./$file\" ; do\n";
         $code .= "${indent}    retries=`expr \$retries - 1`\n";
-        $code .= "${indent}    echo \"Failed to stash file '$file', wait $delay seconds and try again ($retries times left).\"\n";
-        $code .= "${indent}    sleep $delay\n";
+        $code .= "${indent}    echo \"Failed to stash file '$file', wait $retryDelay seconds and try again (\$retries times left).\"\n";
+        $code .= "${indent}    sleep $retryDelay\n";
         $code .= "${indent}  done\n";
         $code .= "${indent}  if [ \$retries -eq 0 ] ; then\n";
-        $code .= "${indent}    echo Failed to stash file '$file', removing incomplete copy.\n";
+        $code .= "${indent}    echo \"Failed to stash file '$file', removing incomplete copy in '$pr:$ns/$path/$file'.\"\n";
         $code .= "${indent}    $dx rm --recursive \"$pr:$ns/$path/$file\"\n";
         $code .= "${indent}    exit 1\n";
         $code .= "${indent}  fi\n";
@@ -397,7 +404,17 @@ sub stashSeqStore ($) {
     return   if (! -e "$asm.seqStore/info");
 
     if (defined(isOS())) {
-        #  Tar up the store metadata, then upload.
+
+        #  Stash the individual data files.  If they exist in the store, don't restash.
+
+        for (my $bIdx="0000"; (-e "./$asm.seqStore/blobs.$bIdx"); $bIdx++) {
+            if (! fileExists("$asm.seqStore/blobs.$bIdx", 1)) {
+                stashFile("$asm.seqStore/blobs.$bIdx");
+            }
+        }
+
+        #  Tar up the store metadata, then upload.  If we failed to upload any blob file,
+        #  we won't upload the metadata, and the store will need to be regenerated.
 
         $files  = "./$asm.seqStore.err ";
         $files .= "./$asm.seqStore.ssi ";
@@ -419,14 +436,6 @@ sub stashSeqStore ($) {
         stashFile("$asm.seqStore.tar.gz");
 
         unlink "./$asm.seqStore.tar.gz";
-
-        #  Now stash the individual data files.  If they exist in the store, don't restash.
-
-        for (my $bIdx="0000"; (-e "./$asm.seqStore/blobs.$bIdx"); $bIdx++) {
-            if (! fileExists("$asm.seqStore/blobs.$bIdx", 1)) {
-                stashFile("$asm.seqStore/blobs.$bIdx");
-            }
-        }
     }
 }
 
@@ -438,9 +447,9 @@ sub fetchSeqStore ($) {
     return   if (! fileExists("$asm.seqStore.tar.gz", 1));
 
     if (defined(isOS())) {
-        print STDERR "\n"                                                  if ($showWork);
-        print STDERR "fetchStore()-- Retrieving store '$asm.seqStore'\n"   if ($showWork);
-        print STDERR "\n"                                                  if ($showWork);
+        print STDERR "--\n"                                                   if ($showWork);
+        print STDERR "-- Retrieving '$asm.seqStore' from object storage.\n"   if ($showWork);
+        print STDERR "--\n"                                                   if ($showWork);
 
         fetchFile("$asm.seqStore.tar.gz");
 
@@ -465,7 +474,7 @@ sub fetchSeqStoreShellCode ($$$) {
         $code .= "\n";
         $code .= "${indent}if [ ! -e $root/$asm.seqStore/info ] ; then\n";
         $code .= "${indent}  echo \"\"\n";
-        $code .= "${indent}  echo In '`pwd`', fetching '$asm.seqStore' into '$root'.\n";
+        $code .= "${indent}  echo \"In '`pwd`', fetching '$asm.seqStore' into '$root'.\"\n";
         $code .= "${indent}  echo \"\"\n";
         $code .= "${indent}  cd $root\n";
         $code .= fetchFileShellCode(".", "$asm.seqStore.tar.gz", "${indent}  ");
@@ -559,7 +568,19 @@ sub stashOvlStore ($$) {
     return   if (! -e "./$base/$asm.ovlStore/index");
 
     if (defined(isOS())) {
-        print STDERR "stashOvlStore()-- Saving overlap store '$base/$asm.ovlStore'\n"   if ($showWork);
+        print STDERR "--\n"                                                   if ($showWork);
+        print STDERR "-- Saving '$base/$asm.ovlStore' to object storage.\n"   if ($showWork);
+        print STDERR "--\n"                                                   if ($showWork);
+
+        #  Stash data files.
+
+        for (my $bIdx="0001"; (-e "./$base/$asm.ovlStore/$bIdx<001>");   $bIdx++) {
+        for (my $sIdx="001";  (-e "./$base/$asm.ovlStore/$bIdx<$sIdx>"); $sIdx++) {
+            if (! fileExists("$base/$asm.ovlStore/$bIdx<$sIdx>", 1)) {
+                stashFile("$base/$asm.ovlStore/$bIdx<$sIdx>");
+            }
+        }
+        }
 
         #  Stash the store metadata.
 
@@ -576,16 +597,6 @@ sub stashOvlStore ($$) {
 
         stashFile("$base/$asm.ovlStore.tar.gz");
         unlink "$base/$asm.ovlStore.tar.gz";
-
-        #  Stash data files.
-
-        for (my $bIdx="0001"; (-e "./$base/$asm.ovlStore/$bIdx<001>");   $bIdx++) {
-        for (my $sIdx="001";  (-e "./$base/$asm.ovlStore/$bIdx<$sIdx>"); $sIdx++) {
-            if (! fileExists("$base/$asm.ovlStore/$bIdx<$sIdx>", 1)) {
-                stashFile("$base/$asm.ovlStore/$bIdx<$sIdx>");
-            }
-        }
-        }
     }
 }
 
@@ -604,6 +615,14 @@ sub stashOvlStoreShellCode ($$) {
         $code .= "fi\n";
         $code .= "\n";
         $code .= "#\n";
+        $code .= "#  Upload data files.\n";
+        $code .= "#\n";
+        $code .= "\n";
+        $code .= "for ff in `ls $asm.ovlStore/????\\<???\\>` ; do\n";
+        $code .= stashFileShellCode("$base", "\$ff", "  ");
+        $code .= "done\n";
+        $code .= "\n";
+        $code .= "#\n";
         $code .= "#  Upload the metadata files.  These shouldn't exist, so we don't bother trying to remove before uploading.\n";
         $code .= "#\n";
         $code .= "\n";
@@ -617,13 +636,6 @@ sub stashOvlStoreShellCode ($$) {
         $code .= "gzip -1c > $base/$asm.ovlStore.tar.gz\n";
         $code .= "\n";
         $code .= stashFileShellCode("$base", "$asm.ovlStore.tar.gz", "");
-        $code .= "#\n";
-        $code .= "#  Upload data files.\n";
-        $code .= "#\n";
-        $code .= "\n";
-        $code .= "for ff in `ls $asm.ovlStore/????\\<???\\>` ; do\n";
-        $code .= stashFileShellCode("$base", "\$ff", "  ");
-        $code .= "done\n";
         $code .= "\n";
     }
 
@@ -640,7 +652,8 @@ sub fetchOvlStore ($$) {
     return   if (! fileExists("$base/$asm.ovlStore.tar.gz"));
 
     if (defined(isOS())) {
-        print STDERR "fetchStore()-- Retrieving store '$base/$asm.ovlStore'\n"   if ($showWork);
+        print STDERR "--\n"                                                         if ($showWork);
+        print STDERR "-- Retrieving '$base/$asm.ovlStore' from object storage.\n"   if ($showWork);
 
         fetchFile("$base/$asm.ovlStore.tar.gz");
 
@@ -649,6 +662,8 @@ sub fetchOvlStore ($$) {
         }
 
         unlink "$base/$asm.ovlStore.tar.gz";
+
+        print STDERR "--\n"                                                         if ($showWork);
     }
 }
 
@@ -664,7 +679,7 @@ sub fetchOvlStoreShellCode ($$$) {
     if (defined(isOS())) {
         $code .= "\n";
         $code .= "${indent}if [ ! -e $root/$base/$asm.ovlStore/index ] ; then\n";
-        $code .= "${indent}  echo In `pwd`, fetching '$base' / '$asm.ovlStore.tar.gz', unzipping in '$root/$base'\n";
+        $code .= "${indent}  echo \"In `pwd`, fetching '$base' / '$asm.ovlStore.tar.gz', unzipping in '$root/$base'.\"\n";
         $code .= fetchFileShellCode($base, "$asm.ovlStore.tar.gz", "${indent}  ");
         $code .= "${indent}  gzip -dc $asm.ovlStore.tar.gz | tar -C $root/$base -xf -\n";
         $code .= "${indent}  rm -f $asm.ovlStore.tar.gz\n";
@@ -716,7 +731,7 @@ sub fetchTigStoreShellCode ($$$$$) {
     if (defined(isOS())) {
         $code .= "\n";
         $code .= "${indent}if [ ! -e $root/$base/$asm.$type/seqDB.v$vers.dat ] ; then\n";
-        $code .= "${indent}  echo In `pwd`, fetching '$base' / '$asm.$type/seqDBv$vers'\n";
+        $code .= "${indent}  echo \"In `pwd`, fetching '$base' / '$asm.$type/seqDBv$vers'/\"\n";
         $code .= "${indent}  mkdir -p $root/$base/$asm.$type\n";
         $code .= "${indent}  cd $root/$base/$asm.$type\n";
         $code .= fetchFileShellCode("$base/$asm.$type", "seqDB.v$vers.dat", "${indent}  ");
@@ -749,11 +764,7 @@ sub stashMerylShellCode ($$$) {
     if (defined(isOS())) {
         $code .= "\n";
         $code .= "${indent}if [ -e ./$name/merylIndex ] ; then\n";
-        $code .= "${indent}  echo In `pwd`, Storing '$path' '$name'\n";
-        $code .= "\n";
-        $code .= "${indent}  tar -cf - ./$name/*Index | gzip -1c > ./$name.tar.gz\n";
-        $code .= stashFileShellCode($path, "$name.tar.gz", "${indent}  ");
-        $code .= "${indent}  rm -f ./$name.tar.gz\n";
+        $code .= "${indent}  echo \"In `pwd`, storing '$path' '$name'.\"\n";
         $code .= "\n";
         $code .= "${indent}  cd $name\n";
         $code .= "\n";
@@ -767,7 +778,12 @@ sub stashMerylShellCode ($$$) {
         $code .= "${indent}            0x111000 0x111001 0x111010 0x111011 0x111100 0x111101 0x111110 0x111111 ; do\n";
         $code .= stashFileShellCode("$path/$name", "\$ff.merylData", "${indent}    ");
         $code .= "${indent}  done\n";
+        $code .= "\n";
         $code .= "${indent}  cd -\n";
+        $code .= "\n";
+        $code .= "${indent}  tar -cf - ./$name/*Index | gzip -1c > ./$name.tar.gz\n";
+        $code .= stashFileShellCode($path, "$name.tar.gz", "${indent}  ");
+        $code .= "${indent}  rm -f ./$name.tar.gz\n";
         $code .= "${indent}fi\n";
     }
 
@@ -785,7 +801,7 @@ sub fetchMerylShellCode ($$$) {
     if (defined(isOS())) {
         $code .= "\n";
         $code .= "${indent}if [ ! -e ./$name/merylIndex ] ; then\n";
-        $code .= "${indent}  echo In `pwd`, fetching '$path' '$name.tar.gz'\n";
+        $code .= "${indent}  echo \"In `pwd`, fetching '$path' '$name.tar.gz'.\"\n";
         $code .= "\n";
         $code .= fetchFileShellCode($path, "$name.tar.gz", "${indent}  ");
         $code .= "${indent}  gzip -dc ./$name.tar.gz | tar -xf -\n";
